# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ouro API
  version: 0.3.0
  description: |
    The Self-Evolving Signal System API.
    
    Ouro is a 7-layer signal processing pipeline that captures human creative signals,
    understands intent, executes optimal tool chains, builds artifacts, recovers patterns,
    and evolves autonomously.
    
    ## Authentication
    Optional. If OURO_API_KEY is set, include it as X-Api-Key header.
    
    ## Core Concepts
    - **Signal**: Any input from a human (text, voice, image, video, file)
    - **Intent**: What the human wants (create, modify, explore, capture, connect, compose)
    - **Execution Plan**: A DAG of tool invocations
    - **Artifact**: A produced output (code, document, image, etc.)
    - **Pattern**: A recovered behavioral signal (creativity trigger, domain preference, etc.)
    - **Evolution**: A system self-improvement event
servers:
  - url: http://localhost:3001
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
...
paths:
...
  /api/prompts/{name}:
    get:
      summary: Get active prompt template
      tags: [Prompts]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Active prompt content
    post:
      summary: Create new prompt version
      tags: [Prompts]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                variables:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: New version created
...